Fix build with latest Inform6-src
C99's `vsnprintf` is now required so we have to include a reimplementation. This change also appears to require a conversion from a double (64-bit float) to a single, and GCC's/PSION's ARM library doesn't have a native implementation of this, therefore we have to stub it out for now to get binaries to build. I need to test if this runs on real hardware now.